% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumany.R
\name{cumany}
\alias{cumany}
\alias{cumall}
\title{Cumulative versions of any() and all()}
\usage{
cumany(x)

cumall(x)
}
\arguments{
\item{x}{A logical vector}
}
\value{
A logical vector the same size as \code{x}
}
\description{
Cumulative versions of any() and all()
}
