% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-package.R
\docType{package}
\name{funs-package}
\alias{funs}
\alias{funs-package}
\title{funs: Useful Vectorised Functions}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
