% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_first.R, R/grouped_max.R,
%   R/grouped_mean.R, R/grouped_min.R, R/grouped_sd.R, R/grouped_sum.R,
%   R/grouped_var.R
\name{grouped_first}
\alias{grouped_first}
\alias{grouped_last}
\alias{grouped_nth}
\alias{grouped_max}
\alias{grouped_mean}
\alias{grouped_min}
\alias{grouped_sd}
\alias{grouped_sum}
\alias{grouped_var}
\title{grouped means, ...}
\usage{
grouped_first(x, default = NA)

grouped_last(x, default = NA)

grouped_nth(x, n, default = NA)

grouped_max(x, na.rm = FALSE)

grouped_mean(x, na.rm = FALSE)

grouped_min(x, na.rm = FALSE)

grouped_sd(x, na.rm = FALSE)

grouped_sum(x, na.rm = FALSE)

grouped_var(x, na.rm = FALSE)
}
\arguments{
\item{x}{list of vectors}

\item{na.rm}{Should NA values be removed ?}
}
\value{
List of means
}
\description{
grouped means, ...
}
\examples{
xs <- vctrs::new_list_of(vctrs::vec_chop(c(1, 2, 3, 4), list(1:2, 3:4)), numeric())

grouped_mean(xs)

grouped_sum(xs)
grouped_var(xs)
grouped_sd(xs)

}
