% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_lead.R
\name{grouped_lead}
\alias{grouped_lead}
\alias{grouped_lag}
\title{grouped lead}
\usage{
grouped_lead(x, n, default = NA)

grouped_lag(x, n, default = NA)
}
\arguments{
\item{x}{list of vectors}

\item{n}{number of positions to lead by}

\item{default}{Value for non existing rows

xs <- vctrs::new_list_of(vctrs::vec_chop(c(1, 2, 3, 4), list(1:2, 3:4)), numeric())

grouped_lead(xs)}
}
\description{
grouped lead
}
