% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_mean.R
\name{grouped_mean}
\alias{grouped_mean}
\title{grouped means}
\usage{
grouped_mean(x, na.rm = FALSE)
}
\arguments{
\item{x}{list of vectors}

\item{na.rm}{Should NA values be removed}
}
\value{
List of means for
}
\description{
grouped means
}
\examples{
xs <- vctrs::new_list_of(vctrs::vec_chop(c(1, 2, 3, 4), list(1:2, 3:4)), numeric())
grouped_mean(xs)

}
