% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incremental.R
\name{incremental_any}
\alias{incremental_any}
\alias{incremental_all}
\alias{incremental_sum}
\alias{incremental_prod}
\alias{incremental_min}
\alias{incremental_max}
\title{Incremental summaries}
\usage{
incremental_any(x)

incremental_all(x)

incremental_sum(x)

incremental_prod(x)

incremental_min(x)

incremental_max(x)
}
\arguments{
\item{x}{for \code{incremental_any()} and \code{incremental_all()} A logical vector}
}
\value{
A vector the same size as \code{x}
\itemize{
\item For \code{incremental_any()} and \code{incremental_all()} : a logical vector
}
}
\description{
Incremental summaries
}
