% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_na.R
\name{is_na}
\alias{is_na}
\title{Are values missing?}
\usage{
is_na(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A logical vector with same size as x.
}
\description{
Returns \code{TRUE} or \code{FALSE} depending on whether the value in \code{x} are missing.
Compared to \code{is.na()}, it follows the tidyverse size principles, so that
\code{length(is_na(df))} equals \code{nrow(df)}.
}
\examples{
is_na(c(NA, 1, NA))

# A row of a data frame is missing if all its values are missing.
df <- data.frame(x = c(NA, 2, 3), y = c(NA, NA, 3))
is_na(df)
}
