% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_summarise.R
\name{jog}
\alias{jog}
\title{For testing purpuses for now}
\usage{
jog(.data, quo)
}
\description{
For testing purpuses for now
}
\examples{
df <- dplyr::group_by(data.frame(x = 1:4, y = c(1,1,2,2)), y)
df \%>\% jog(mean(x))
df \%>\% jog(sum(x))
df \%>\% jog(var(x))
df \%>\% jog(sd(x))

df \%>\% jog(first(x))
}
