% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_if.R
\name{na_if}
\alias{na_if}
\title{Convert values to NA}
\usage{
na_if(x, y)
}
\arguments{
\item{x}{A vector to modify}

\item{y}{Values to replace with \code{NA}, or a predicate function}
}
\description{
Convert values to NA
}
\examples{
x <- 1:10

# turn specific values to NA
na_if(x, c(6, 8))

# or use a predicate
na_if(x, ~ .x > 5)

df <- data.frame(x = 1:10, y = 1:10)
na_if(df, ~ .x$y > 8)

}
\seealso{
\code{\link[vctrs:vec_match]{vctrs::vec_in()}} is used when \code{y} is a vector
}
