% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample}
\alias{sample}
\title{Take a random sample from a vector.}
\usage{
sample(x, size = vec_size(x), replace = FALSE, w = NULL)
}
\arguments{
\item{x}{A vector.}

\item{size}{Number of samples to take, Defaults to the size of \code{x},
generating a random permutation.}

\item{replace}{Should samples be taken with or without replacement?}

\item{w}{Optionally, a vector of weights. If provided should be a
numeric vector the same size as x.}
}
\description{
Take a random sample from a vector.
}
\examples{
sample(1:5)
sample(1:100, 10)
sample(1:10, 100, replace = TRUE)

# samples of data frames are taken row wise, making it easy to
# do bootstrapping
sample(iris, 10, replace = TRUE)

# Unlike base::sample(), this is safe to use with inputs of length 1
sample(3:5)
sample(4:5)
sample(5:5)
base::sample(5:5)
}
