% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_mean.R
\name{grouped_mean}
\alias{grouped_mean}
\title{grouped means}
\usage{
grouped_mean(x, na.rm = FALSE)
}
\description{
grouped means
}
\examples{
xs <- mince(c(1, 2, 3, 4), list(1:2, 3:4))
grouped_mean(xs)

}
