% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_if.R
\name{na_if}
\alias{na_if}
\title{Convert values to NA}
\usage{
na_if(x, y)
}
\arguments{
\item{x}{A vector to modify}

\item{y}{What to replace with \code{NA}. Can be:
\itemize{
\item A vector whose type matches \code{x}, in that case \code{\link[vctrs:vec_match]{vctrs::vec_in()}} is used
to determine which values are replace with \code{NA}.
\item A predicate function applied to \code{x} returning a logical vector, integer or character vector. See \code{\link[vctrs:vec_slice]{vctrs::vec_slice<-()}} for details.
}}
}
\description{
Convert values to NA
}
\examples{
x <- 1:10

# turn specific values to NA
na_if(x, c(6, 8))

# or use a predicate
na_if(x, ~ .x > 5)

df <- data.frame(x = 1:10, y = 1:10)
na_if(df, ~ .x$y > 8)

}
